backend/shipquicker/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.4
importlib-metadata==4.11.3
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
zipp==3.7.0
